addressing code review

+++ azure-mgmt-batchai/src/main/java/com/microsoft/azure/management/batchai/RemoteLoginInformation.java
@@ -33,5 +33,5 @@ public interface RemoteLoginInformation extends
     /**
      * @return port number
      */
-    double port();
+    int port();
 }
